+++ src/error/panic.rs
@@ -78,7 +78,7 @@ mod reexports {
     /// <span class='stab portability' title='re-exported from rust&#39;s `core`'>`core`</span>
     /// Panics the current thread.
     ///
-    #[doc = "*Re-exported from [`core::panic`][macro@panic]*."]
+    #[doc = "*Re-exported from [`core::panic`][macro@crate::_dep::_core::panic]*."]
     #[doc = "\n\n---"]
     #[macro_export]
     macro_rules! panic { ($($tt:tt)*) => { core::panic![$($tt)*] } }